refactor(grading): move CompositeFold helpers to core.grading (Stage 1 of 2)
- New pure module tolokaforge/core/grading/composite_fold.py holds
  StateChecksOutcome, TrialVerdict, resolve_state_checks_component,
  compose_trial_verdict, build_grade_reasons, combine_grade_components
- Delete same from tolokaforge/runner/grading.py (no re-export shim)
- Renames: RunnerTrialVerdict -> TrialVerdict,
  compose_runner_trial_verdict -> compose_trial_verdict
- Update callers: runner/service.py, grader/composite_dispatch.py,
  core/grading/rubric_migration.py (4 sites), key_manifest.py (4 evaluator
  strings), 10 test files (3 shape groups per plan)
- New .importlinter contract composite-fold-purity (strict): forbids
  composite_fold importing from runner, grader, substrate_live
- Behaviour byte-identical; byte-parity 10-pack gate untouched

Refs #1346
